// rent-a-ride-ui/src/app/services/models/vehicle-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface VehicleRequest {
  carModel: string;
  cover?: string;
  description: string;
  id?: number;
  licensePlate: string;
  ownerName: string;
  phone: string;
  price?: string;
  shareable?: boolean;
}